renamed kuzzleSecurityDocument.js -> SecurityDocument.js
